+++ lib/ui/onboarding/OnboardingUpgradePanel.dart
@@ -101,7 +101,7 @@ class _OnboardingUpgradePanelState extends State<OnboardingUpgradePanel> {
                   RoundedButton(
                     label: Localization().getStringEx('panel.onboarding.upgrade.button.upgrade.title', 'Upgrade'),
                     hint: Localization().getStringEx('panel.onboarding.upgrade.button.upgrade.hint', ''),
-                    textColor: Styles().colors!.white,
+                    textStyle: Styles().textStyles?.getTextStyle("widget.colourful_button.title.large.accent"),
                     borderColor: Styles().colors!.fillColorSecondary,
                     backgroundColor: Styles().colors!.fillColorSecondary,
                     onTap: () => _onUpgradeClicked(context),