+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Export button text styles to styles.json [#2932](https://github.com/rokwire/illinois-app/issues/2932).
 
 ## [5.0.8] - 2023-06-12
 ### Added